+++ pywren_ibm_cloud/executor.py
@@ -45,7 +45,6 @@ class FunctionExecutor:
 
         :return `FunctionExecutor` object.
         """
-        self.start_time = time.time()
         self._state = FunctionExecutor.State.New
         self.is_pywren_function = is_pywren_function()
 
@@ -417,7 +416,7 @@ class FunctionExecutor:
 
         return result
 
-    def create_execution_plots(self, dst_dir, dst_file_name, fs=None):
+    def plot(self, fs=None, dst=None):
         """
         Creates timeline and histogram of the current execution in dst_dir.
 
@@ -436,20 +435,14 @@ class FunctionExecutor:
             logger.debug('ExecutorID {} - No futures ready to plot'.format(self.executor_id))
             return
 
-        start_time = self.start_time if not fs else ftrs_to_plot[0]._call_metadata['host_submit_time']-1
-
         logging.getLogger('matplotlib').setLevel(logging.WARNING)
         from pywren_ibm_cloud.plots import create_timeline, create_histogram
 
         msg = 'ExecutorID {} - Creating execution plots'.format(self.executor_id)
         print(msg) if not self.log_level else logger.info(msg)
 
-        call_status = [f._call_status for f in ftrs_to_plot]
-        call_metadata = [f._call_metadata for f in ftrs_to_plot]
-
-        dst_dir = os.path.expanduser(dst_dir)
-        create_timeline(dst_dir, dst_file_name, start_time, call_status, call_metadata, self.config['ibm_cos'])
-        create_histogram(dst_dir, dst_file_name, start_time, call_status, self.config['ibm_cos'])
+        create_timeline(ftrs_to_plot, dst)
+        create_histogram(ftrs_to_plot, dst)
 
     def clean(self, fs=None, local_execution=True):
         """